backend/apps/accounts/urls.py
"""Auth and accounts API URLs."""
from django.urls import path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path("login/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    # TODO: signup, logout, verify-email, verify-phone, me, change-password
]
