backend/requirements.txt
# ===== Django Core =====
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2

# ===== Database =====
psycopg[binary]==3.1.19
dj-database-url==2.2.0

# ===== Auth =====
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.3
PyJWT==2.8.0

# ===== Async / Tasks =====
celery==5.4.0
redis==5.0.4
django-celery-beat==2.6.0
django-redis==5.4.0

# ===== API Docs =====
drf-spectacular==0.27.2

# ===== Storage =====
django-storages==1.14.3
boto3==1.34.100

# ===== Utilities =====
Pillow==10.3.0
python-dateutil==2.9.0
phonenumbers==8.13.37
django-phonenumber-field==7.3.0
whitenoise==6.6.0

# ===== Observability =====
sentry-sdk==2.2.0
structlog==24.1.0

# ===== Security =====
cryptography==42.0.7
argon2-cffi==23.1.0

# ===== Payments =====
razorpay==1.4.2
stripe==9.5.0

# ===== Development / Testing =====
pytest==8.2.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.2.0
ipython==8.24.0
ruff==0.4.4
mypy==1.10.0
django-stubs==5.0.0
